requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
anyio==4.8.0
    # via
    #   httpx
    #   openai
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
cytoolz==1.0.1
    # via eth-utils
distro==1.9.0
    # via openai
eth-abi==4.2.1
    # via python-cartesi
eth-hash==0.7.1
    # via eth-utils
eth-typing==5.1.0
    # via
    #   eth-abi
    #   eth-utils
eth-utils==5.2.0
    # via eth-abi
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jiter==0.8.2
    # via openai
openai==1.62.0
    # via -r requirements.in
packaging==24.2
    # via pytest
parsimonious==0.9.0
    # via eth-abi
pluggy==1.5.0
    # via pytest
pycryptodome==3.19.0
    # via
    #   -r requirements.in
    #   python-cartesi
pydantic==1.10.21
    # via
    #   openai
    #   python-cartesi
pytest==7.4.4
    # via
    #   -r requirements.in
    #   python-cartesi
python-cartesi==0.1.2
    # via -r requirements.in
regex==2024.11.6
    # via parsimonious
requests==2.32.3
    # via
    #   -r requirements.in
    #   python-cartesi
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tomli==2.2.1
    # via pytest
toolz==1.0.0
    # via cytoolz
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   eth-typing
    #   openai
    #   pydantic
urllib3==2.3.0
    # via requests
